+++ abstractions/typescript/src/serialization/parsable.ts
@@ -9,7 +9,7 @@ export interface Parsable {
    * Gets the deserialization information for this object.
    * @return The deserialization information for this object where each entry is a property key with its deserialization callback.
    */
-  getFieldDeserializers<T>(): Map<string, (item: T, node: ParseNode) => void>;
+  getFieldDeserializers<T>(): Record<string, (item: T, node: ParseNode) => void>;
   /**
    * Writes the objects properties to the current writer.
    * @param writer The writer to write to.